:bug: Fix package regex matching
Fix an issue where similarly named packages collide.
For example, imagine this `pkgify` config:

```json
"pkgify": {
  "packages": {
    "some-package": "path/to/some-package"
  }
}
```

And then you also install `some-package-plugin` in your
`dependencies`. When a `require` statement for `some-package-plugin`
goes through `pkgify`, it currently matches `some-package` which
is undesirable since `some-package` and `some-package-plugin` are
completely different packages. To work around this, test whether
the `some-package` is `require`d directly via `^...$` *or* if something
*within* `some-package` is `require`d via `^.../`, ie
`some-package/something-in-it`.